chore(Rendering): clean some common code
Move common code from api specific render windows into
a common superclass. A prep stage to making a more
general hardware selector api.
