+++ Sources/Rendering/WebGPU/RenderWindow/index.js
@@ -5,7 +5,7 @@ import vtkWebGPUDevice from 'vtk.js/Sources/Rendering/WebGPU/Device';
 import vtkWebGPUSwapChain from 'vtk.js/Sources/Rendering/WebGPU/SwapChain';
 import vtkWebGPUViewNodeFactory from 'vtk.js/Sources/Rendering/WebGPU/ViewNodeFactory';
 import vtkRenderPass from 'vtk.js/Sources/Rendering/SceneGraph/RenderPass';
-import vtkViewNode from 'vtk.js/Sources/Rendering/SceneGraph/ViewNode';
+import vtkRenderWindowViewNode from 'vtk.js/Sources/Rendering/SceneGraph/RenderWindowViewNode';
 // import vtkOpenGLTextureUnitManager from 'vtk.js/Sources/Rendering/OpenGL/TextureUnitManager';
 // import { VtkDataTypes } from 'vtk.js/Sources/Common/Core/DataArray/Constants';
 
@@ -156,132 +156,7 @@ function vtkWebGPURenderWindow(publicAPI, model) {
     return model.containerSize || model.size;
   };
 
-  publicAPI.getFramebufferSize = () => {
-    if (model.activeFramebuffer) {
-      return model.activeFramebuffer.getSize();
-    }
-    return model.size;
-  };
-
-  publicAPI.isInViewport = (x, y, viewport) => {
-    const vCoords = viewport.getViewportByReference();
-    const size = publicAPI.getFramebufferSize();
-    if (
-      vCoords[0] * size[0] <= x &&
-      vCoords[2] * size[0] >= x &&
-      vCoords[1] * size[1] <= y &&
-      vCoords[3] * size[1] >= y
-    ) {
-      return true;
-    }
-    return false;
-  };
-
-  publicAPI.getViewportSize = (viewport) => {
-    const vCoords = viewport.getViewportByReference();
-    const size = publicAPI.getFramebufferSize();
-
-    return [
-      (vCoords[2] - vCoords[0]) * size[0],
-      (vCoords[3] - vCoords[1]) * size[1],
-    ];
-  };
-
-  publicAPI.getViewportCenter = (viewport) => {
-    const size = publicAPI.getViewportSize(viewport);
-    return [size[0] * 0.5, size[1] * 0.5];
-  };
-
-  publicAPI.displayToNormalizedDisplay = (x, y, z) => {
-    const size = publicAPI.getFramebufferSize();
-    return [x / size[0], y / size[1], z];
-  };
-
-  publicAPI.normalizedDisplayToDisplay = (x, y, z) => {
-    const size = publicAPI.getFramebufferSize();
-    return [x * size[0], y * size[1], z];
-  };
-
-  publicAPI.worldToView = (x, y, z, renderer) => renderer.worldToView(x, y, z);
-
-  publicAPI.viewToWorld = (x, y, z, renderer) => renderer.viewToWorld(x, y, z);
-
-  publicAPI.worldToDisplay = (x, y, z, renderer) => {
-    const val = renderer.worldToView(x, y, z);
-    const dims = publicAPI.getViewportSize(renderer);
-    const val2 = renderer.viewToProjection(
-      val[0],
-      val[1],
-      val[2],
-      dims[0] / dims[1]
-    );
-    const val3 = renderer.projectionToNormalizedDisplay(
-      val2[0],
-      val2[1],
-      val2[2]
-    );
-    return publicAPI.normalizedDisplayToDisplay(val3[0], val3[1], val3[2]);
-  };
-
-  publicAPI.displayToWorld = (x, y, z, renderer) => {
-    const val = publicAPI.displayToNormalizedDisplay(x, y, z);
-    const val2 = renderer.normalizedDisplayToProjection(val[0], val[1], val[2]);
-    const dims = publicAPI.getViewportSize(renderer);
-    const val3 = renderer.projectionToView(
-      val2[0],
-      val2[1],
-      val2[2],
-      dims[0] / dims[1]
-    );
-    return renderer.viewToWorld(val3[0], val3[1], val3[2]);
-  };
-
-  publicAPI.normalizedDisplayToViewport = (x, y, z, renderer) => {
-    let vCoords = renderer.getViewportByReference();
-    vCoords = publicAPI.normalizedDisplayToDisplay(vCoords[0], vCoords[1], 0.0);
-    const coords = publicAPI.normalizedDisplayToDisplay(x, y, z);
-    return [coords[0] - vCoords[0] - 0.5, coords[1] - vCoords[1] - 0.5, z];
-  };
-
-  publicAPI.viewportToNormalizedViewport = (x, y, z, renderer) => {
-    const size = publicAPI.getViewportSize(renderer);
-    if (size && size[0] !== 0 && size[1] !== 0) {
-      return [x / (size[0] - 1.0), y / (size[1] - 1.0), z];
-    }
-    return [x, y, z];
-  };
-
-  publicAPI.normalizedViewportToViewport = (x, y, z) => {
-    const size = publicAPI.getFramebufferSize();
-    return [x * (size[0] - 1.0), y * (size[1] - 1.0), z];
-  };
-
-  publicAPI.displayToLocalDisplay = (x, y, z) => {
-    const size = publicAPI.getFramebufferSize();
-    return [x, size[1] - y - 1, z];
-  };
-
-  publicAPI.viewportToNormalizedDisplay = (x, y, z, renderer) => {
-    let vCoords = renderer.getViewportByReference();
-    vCoords = publicAPI.normalizedDisplayToDisplay(vCoords[0], vCoords[1], 0.0);
-    const x2 = x + vCoords[0] + 0.5;
-    const y2 = y + vCoords[1] + 0.5;
-    return publicAPI.displayToNormalizedDisplay(x2, y2, z);
-  };
-
-  publicAPI.getPixelData = (x1, y1, x2, y2) => {
-    const pixels = new Uint8Array((x2 - x1 + 1) * (y2 - y1 + 1) * 4);
-    model.context.readPixels(
-      x1,
-      y1,
-      x2 - x1 + 1,
-      y2 - y1 + 1,
-      model.context.RGBA,
-      model.context.UNSIGNED_BYTE,
-      pixels
-    );
-    return pixels;
-  };
+  publicAPI.getFramebufferSize = () => model.size;
 
   publicAPI.create3DContext = () => {
     model.initializing = true;
@@ -449,7 +324,6 @@ const DEFAULT_VALUES = {
   swapChain: null,
   cullFaceEnabled: false,
   depthMaskEnabled: true,
-  size: [300, 300],
   cursorVisibility: true,
   cursor: 'pointer',
   textureUnitManager: null,
@@ -457,7 +331,6 @@ const DEFAULT_VALUES = {
   containerSize: null,
   renderPasses: [],
   notifyStartCaptureImage: false,
-  activeFramebuffer: null,
   imageFormat: 'image/png',
   useOffScreen: false,
   useBackgroundImage: false,
@@ -484,7 +357,7 @@ export function extend(publicAPI, model, initialValues = {}) {
   model.swapChain = vtkWebGPUSwapChain.newInstance();
 
   // Inheritance
-  vtkViewNode.extend(publicAPI, model, initialValues);
+  vtkRenderWindowViewNode.extend(publicAPI, model, initialValues);
 
   model.myFactory = vtkWebGPUViewNodeFactory.newInstance();
   /* eslint-disable no-use-before-define */
@@ -514,11 +387,6 @@ export function extend(publicAPI, model, initialValues = {}) {
     'notifyStartCaptureImage',
     'cursor',
     'useOffScreen',
-    // might want to make this not call modified as
-    // we change the active framebuffer a lot. Or maybe
-    // only mark modified if the size or depth
-    // of the buffer has changed
-    'activeFramebuffer',
   ]);
 
   macro.setGetArray(publicAPI, model, ['size'], 2);
